server/requirements.txt
# Text extraction libraries
pypdf==4.2.0
python-docx==1.1.2
chardet==5.2.0

# Web framework
Flask==3.0.0
Flask-CORS==4.0.0

# Azure Storage
azure-storage-blob==12.19.0

# File handling
Werkzeug==3.0.1

# Development
python-dotenv==1.0.0
